Restructure card views
