+++ Swanki/View/CardSide.swift
@@ -1,0 +1,8 @@
+// Copyright © 2019-present Brian Dewey.
+
+import Foundation
+
+public enum CardSide {
+  case front
+  case back
+}
