+++ Swanki/View/Cards/CardQuizView.swift
@@ -5,20 +5,23 @@ import SpacedRepetitionScheduler
 import SwiftData
 import SwiftUI
 
+/// Quizzes a learner on the contents of a specific card.
 struct CardQuizView: View {
+  /// The card to quiz.
   var card: Card
+
+  /// A closure invoked when the learner selects an answer.
   var didSelectAnswer: ((CardAnswer, SpacedRepetitionScheduler.Item, TimeInterval) -> Void)?
 
   @State private var viewDidAppearTime: Date?
   @State private var isShowingBack = false
 
   var body: some View {
-    Self._printChanges()
-    return VStack {
-      NewCardView(card: card, showFront: true)
+    VStack {
+      GenericCardView(card: card, cardSide: .front)
       if isShowingBack {
         Divider()
-        NewCardView(card: card, showFront: false)
+        GenericCardView(card: card, cardSide: .back)
         CardAnswerButtonRow(answers: possibleAnswers) { answer, item in
           let duration = viewDidAppearTime.flatMap { Date.now.timeIntervalSince($0) } ?? 2
           didSelectAnswer?(answer, item, duration)
@@ -42,59 +45,6 @@ struct CardQuizView: View {
 
   private var possibleAnswers: [(key: CardAnswer, value: SpacedRepetitionScheduler.Item)] {
     SpacedRepetitionScheduler.builtin.scheduleItem(.init(card))
-  }
-
-  private var front: String {
-    card.note?.field(at: 0) ?? ""
-  }
-
-  private var back: String {
-    card.note?.field(at: 1) ?? ""
-  }
-}
-
-public enum CardType: String, Codable {
-  case frontThenBack
-  case backThenFront
-}
-
-struct NewCardView: View {
-  var card: Card
-  var showFront: Bool
-
-  var body: some View {
-    switch card.type {
-    case .frontThenBack:
-      FrontThenBackCard(card: card, showFront: showFront)
-    case .backThenFront:
-      BackThenFrontCard(card: card, showFront: showFront)
-    }
-  }
-}
-
-struct FrontThenBackCard: View {
-  var card: Card
-  var showFront: Bool
-
-  var body: some View {
-    if showFront {
-      Text(card.note?.field(at: 0) ?? "")
-    } else {
-      Text(card.note?.field(at: 1) ?? "")
-    }
-  }
-}
-
-struct BackThenFrontCard: View {
-  var card: Card
-  var showFront: Bool
-
-  var body: some View {
-    if showFront {
-      Text(card.note?.field(at: 1) ?? "")
-    } else {
-      Text(card.note?.field(at: 0) ?? "")
-    }
   }
 }
